Run npm test in CI. Run dotenv. Rename util/ to utils/
