+++ .github/workflows/nodejs.yml
@@ -26,5 +26,6 @@ jobs:
         run: |
           npm ci
           npm run build --if-present
+          npm test
         env:
           CI: true
