+++ src/config/index.ts
@@ -1,3 +1,6 @@
+import dotenv from 'dotenv';
+dotenv.config();
+
 const config = {
   nodeEnv: process.env.NODE_ENV || 'development',
 };
